Simple implementation to dynamically load and execute kernel from Python with pybind11;

+++ execute/examples/dynamic-load.cpp
@@ -6,7 +6,7 @@
 extern "C" {
   void foo(size_t num_elements, void *inputs, double *outputs) {
     std::cout << "Dynamically loaded function foo" << std::endl;
-    int *int_inputs = (int *) inputs;
+    double *int_inputs = (double *) inputs;
     for (int i = 0; i < num_elements; ++i) {
       outputs[i] = int_inputs[i] * 2.0;
     }
